+++ src/index.ts
@@ -1,4 +1,11 @@
-export { id, locateAttr, locateNode, locate, locateID } from "./locate";
+export {
+	locateAttr,
+	locateNode,
+	locate,
+	locateID,
+	getID,
+	setID,
+} from "./locate";
 export { delegateClick, delegateEvent } from "./delegateClick";
 export { clickOutside } from "./clickOutside";
 export { calculatePosition, getAbsParent } from "./positioning";